Use fprintf(stderr,  rather than message(LOG, for page number output
As a result, only message(FATAL, is ever used, so remove the flag and
simplify message and its callers.

Remove embedded newlines from the end of all calls to message, as
message takes care of this.

Thanks to Akira Kakuto and Peter Breitenlohner for the initial report
and a hint towards the solution respectively.