Sort out resizing
The auto-resizing code originally from psresize had been integrated into
pstops.c. This was wrong: if the input paper size and output paper size
were different, the code was automatically activated. But this made it
impossible for psnup to signal that the input paper size was A5, but
already have decided how to scale the pages on to output A4.

In fact, the input paper size is irrelevant (currently) to pstops.

Hence, delete the auto-resizing code.

This prevents psresize from working, but observe that psresize is just
psnup -1, so reimplement psresize as a call to psnup, and use its
resizing code.

Without input paper size specification, pstops had no way to know when
the paper size was changing and hence strip the size headers. Add a -S
argument to signal this.

Update pstops.1 to document the changes.

This change also fixes the psnup test, which was broken by the
translation to Perl.