Split out common routines into module
When pstops is rewritten in Perl, it will share argument parsing routines
and others with psnup. Hence, split those out into a module, and add the
necessary rules to the build system to install the module and make it
available to psnup (and in future other programs that use it) during in-tree
testing.