pstops: allow pagespecs to be omitted (default to “0”; fix #11)
Add --specs argument.

However, remain backwards-compatible, and still try to parse a command line
argument as a spec. This will only go wrong if a file argument is a valid
spec, which is unlikely. Nevertheless, do not document this syntax.

Also ensure all die errors end with a newline so as not to add an (ugly)
Perl error “at line” message.