Add support for PDF files
This adds a parallel set of “pdfutils”, and a parallel set of tests.

Also add a test for “ps/pdfnup --draw”, and fix a bug in psnup where the
“--draw” and “--quiet” arguments were not passed on to pstops.

There is no mention of them in the documentation as yet, but I intend to
merge the PDF support so that one set of scripts can cope with both file
types.