Merge pdf* scripts into ps* scripts
Convert all I/O to binary. This is necessary so we can test the MIME type of
the input before knowing whether it’s text or binary. It doesn’t hurt to
handle PostScript as binary, in any case.

Add MIME type detection with “puremagic” to work out whether we’re dealing
with PostScript or PDF in each case.