Slurp input
This simplifies the code because we don’t need to possibly copy stdin to a
temporary file; instead, present input as a (seekable) BytesIO.

Also, we no longer need chainstream.

This removes the “make_seekable” argument from setup_input_and_output,
paving the way for us to encapsulate it in DocumentTransform.