Extract PsReader class from PsDocumentTransform
The {Ps,Pdf}Reader is then passed to the transform class’s constructor.

Also rename {Ps,Pdf}DocumentTransform to {Ps,Pdf}Transform, for consistency.

Rename variables holding Transform classes from “doc” to “transform”.

As well as making the code easier to maintain, this change will allow us to
set the page size of the input from the PostScript file: we need to be able
to do this in all callers of the transform classes (in particular, in
programs other than pstops).