Fix referring to paper size when output paper unset (fix #35)
This revealed some test results that were not exactly right: pstops was
guessing that the input paper size was the same as the output paper
size (which it wasn’t), and was doing things a bit wrong, for example
imposing a portrait input page on the same size landscape output
page (psnup/20-1-flipped-dimensions). The PDF results were also wrong in
this and another similar case!

Also, by parsing the size information, we get slightly different dimensions
in one test (psresize/20-A3in-A4), so the output is slightly different.