PDF: use correct output paper size when it differs from input size
When the transformation is “simple”, that is, fulfils certain criteria such
as having only one pagespec and no transformation, we avoid rewriting all
the pages of the input, and simply copy them to the output. We were
incorrectly using this code path (which should be faster, and also safer in
the sense that it does not mess with the structure of the input as much) by
not checking whether the input and output paper sizes were the same. Do so!

Add a test to avoid this problem recurring.